refactor: changed the CLI to have different subcommands e.g `resolve` and `install`
This adds a more flexible way to add commands to the binary.
