+++ README.md
@@ -53,12 +53,12 @@ There is no detection of existing packages yet.
 
 ## Features
 
-This is a list of current and planned features of `RIP`, the biggest are listed below:
+This is a list of current features of `RIP`, the biggest are listed below:
 
-* [x] Downloading and aggressive caching of PyPI metadata.
+* [x] Async downloading and aggressive caching of PyPI metadata.
 * [x] Resolving of PyPI packages using [Resolvo](https://github.com/mamba-org/resolvo).
-* [x] Installation of wheel files (see: https://github.com/prefix-dev/rip/issues/6 for last open issues)
-* [x] Support sdist files
+* [x] Installation of wheel files.
+* [x] Support sdist files (must currently adhere to the `PEP 517` and `PEP 518` standards).
 
 More intricacies of the PyPI ecosystem need to be implemented, see our GitHub issues for more details.
